Volleyball
Vladi is a student, lives in Sofia and goes to his hometown from time to time. He is very passionate about volleyball, but is busy on weekdays and plays volleyball only on weekends and holidays. Vladi plays in Sofia every Saturday when he is not at work and does not travel to his hometown, as well as on 2/3 of the holidays. He travels to his hometown h times a year, where he plays volleyball with his old friends on Sundays. Vladi is not at work 3/4 of the weekends he is in Sofia. Separately, during the high years Vladi played 15% more volleyball than normal. We assume that the year has exactly 48 weekends suitable for volleyball.
Write a program that calculates how many times Vladi has played volleyball during the year. Round the result down to the nearest whole number (eg 2.15  2; 9.95 (9).
The input data is entered by the user in the following form:
• The first line contains the word "leap" (leap year) or "normal" (leap year).
• The second line contains the integer p - number of holidays in the year (which are not Saturday and Sunday).
• The third line contains the integer h - the number of weekends in which Vladi travels to his hometown.