Division without remainder
Division without remainder
N-integers in the interval [1… 1000] are given. Of these, some percentage p1 is divisible by 2, another percentage of p2 is divisible by 3, another percentage of p3 is divisible by 4. Write a program that calculates and prints the percentages p1, p2 and p3.
Example: we have n = 10 numbers: 680, 2, 600, 200, 800, 799, 199, 46, 128, 65. We get the following distribution and visualization:
Entrance
On the first line of the input is the integer n (1 ≤ n ≤ 1000) - number of numbers. On the next n lines there is an integer in the interval [1… 1000] - the numbers to be checked by how much they are divisible.
Exit
Print 3 lines on the console, each containing a percentage between 0% and 100%, to two decimal places, for example 25.00%, 66.67%, 57.14%.
• In the first row - the percentage of numbers that are divisible by 2
• On the second line - the percentage of numbers that are divisible by 3
• On the third line - the percentage of numbers that are divisible by 4