Cake
Cake
You are invited to the 30th birthday, when the birthday boy eats a huge cake. However, he does not know how many pieces his guests can take from her. Your task is to write a program that calculates the number of pieces that guests have taken before it ends. You will get the dimensions of the cake in centimeters (width and length - integers in the range [1 ... 1000]) and then on each line, until you receive the command "STOP" or until the cake is finished, the number of pieces that guests take from her. The pieces are square with a size of 1 cm.
Print one of the following lines on the console:
• "{number of pieces} pieces are left." - if you reach STOP and there are pieces of cake left.
• "No more cake left! You need {number of missing pieces} pieces more."