Sums of prime and unprime numbers
Write a program that reads integers from the console in the range of -2,147,483,648 to 2,147,483,647 until a "stop" command is received. Find the sum of all entered primes and the sum of all entered primes. Because, by definition, mathematics, negative numbers cannot be prime, if the following message "Number is negative." In this case, the entered number is ignored and is not added to either of the two amounts, and the program continues its execution, waiting for the next number to be entered.
At the output to print in two lines the two amounts found in the following format:
"Sum of all prime numbers is: {prime numbers sum}"
"Sum of all non prime numbers is: {nonprime numbers sum}"