Equal amounts of even and odd positions
Write a program that reads from the console two six-digit integers in the range of 100,000 to 300,000. Always the first number entered will be less than the second. On the console to print on 1 line separated by a space all the numbers that are between the two numbers read by the console and meet the following condition:
• the sum of the digits of even and odd positions to be equal. If there are no numbers that meet the console condition, no result is displayed.