Hotel room.java
The hotel offers 2 types of rooms: studio and apartment. Write a program that calculates the price for the entire stay for a studio and apartment. Prices depend on the month of stay:
May and October June and September July and August
Studio - BGN 50 / night Studio - BGN 75.20 / night Studio - BGN 76 / night
Apartment - 65 BGN / night Apartment - 68.70 BGN / night Apartment - 77 BGN / night
The following discounts are also available:
• For studio, for more than 7 nights in May and October: 5% discount.
• For studio, for more than 14 nights in May and October: 30% discount.
• For studio, for more than 14 nights in June and September: 20% discount.
• For an apartment, for more than 14 nights, regardless of the month: 10% discount.
Entrance
The input is readable by the console and contains exactly 2 lines entered by the user:
• In the first row is the month - May, June, July, August, September or October
• On the second row is the number of nights - an integer in the range [0 ... 200]
Exit
To print 2 lines on the console:
• On the first line: “Apartment: {price for the whole stay} lv.”
• On the second line: "Studio: {price for the whole stay} lv."
The price for the entire stay is formatted to two decimal places.