In time for the exam.java
A student must go to the exam at a certain time. He arrives at the examination room at a certain time of arrival. It is considered that the student arrived on time if he arrived at the time of the exam or up to half an hour before. If he arrived more than 30 minutes earlier, he was early. If he came after the exam time, he was late. Write a program that reads exam time and arrival time and records whether the student arrived on time, was early or late, and by how many hours or minutes was early or late.
Entrance
The console reads 4 integers (one per line) entered by the user:
• The first line contains the exam time - an integer from 0 to 23.
• The second line contains the minute of the exam - an integer from 0 to 59.
• The third line contains the time of arrival - an integer from 0 to 23.
• The fourth line contains the minute of arrival - an integer from 0 to 59.
Exit
On the first line print:
• “Late” if the student arrives later than the exam time.
• "On time", if the student arrives exactly at the time of the exam or up to 30 minutes earlier.
• “Early” if the student arrives more than 30 minutes before the exam time.
If the student arrives at least one minute apart from the exam time, print on the following line: