Journey.java
Strange, but most people plan their vacation early. A young programmer has a certain budget and free time in a given season. Write a program that will accept the budget and the season at the entrance, and at the exit to earn where the programmer will rest and how much he will spend.
The budget determines the destination, and the season determines how much of the budget you will spend. If it is summer he will rest at the campsite and in the winter at a hotel. If he is in Europe, he will stay in a hotel, regardless of the season. Each campsite or hotel, according to the destination, has its own price that corresponds to a certain percentage of the budget:
• At BGN 100. or less - somewhere in Bulgaria
o Summer - 30% of the budget
o Winter - 70% of the budget
• At BGN 1,000. or a little - somewhere in the Balkans
o Summer - 40% of the budget
o Winter - 80% of the budget
• For more than BGN 1,000. - somewhere in Europe
o When traveling in Europe, regardless of the season will spend 90% of the budget.
Entrance
The input is readable by the console and consists of two lines entered by the user:
• First line - Budget, real number in the range [10.00 ... 5000.00].
• Second row - One of the two possible seasons: "summer" or "winter"
Exit
Two lines must be printed on the console.
• First line - "Somewhere in" [destination] "between" Bulgaria "," Balkans "and" Europe "
• Second line - "{Type of holiday} - {Amount spent}"
o Holidays can be between "Camp" and "Hotel"
o The amount must be rounded to the second decimal place.