Operations between numbers.java
Write a program that reads two integers (N1 and N2) and an operator with which to perform a mathematical operation with them. Possible operations are: Addition (+), Subtraction (-), Multiplication (*), Division (/) and Modular Division (%). When adding, subtracting and multiplying the console, the result must be printed and whether it is even or odd. In ordinary division - the result. In modular division - the remainder. It should be borne in mind that the divisor can be equal to 0 (zero) and not divisible by zero. In this case, a special message must be printed.
Entrance
3 lines entered by the user are read from the console:
• N1 - integer in the range [0 ... 40 000]
• N2 - integer in the range [0 ... 40 000]
• Operator - one character between: "+", "-", "*", "/", "%"
Exit
Print one line on the console:
• If the operation is addition, subtraction or multiplication:
o "{{N1} {operator} {N2} = {result} - {even / odd}"
• If the operation is a division:
o "{N1} / {N2} = {result}" - the result is formatted to the second decimal place
• If the operation is a modular division:
o "{N1}% {N2} = {remainder}"
• In case of division by 0 (zero):
o "Cannot divide {N1} by zero"