Ski vacation.java
Atanas decided to spend his vacation in Bansko and ski. Before he leaves, however, he must book a hotel and calculate how much his stay will cost. There are the following types of premises, with the following prices for accommodation:
 "room for one person" - BGN 18.00 per night
 "apartment" - BGN 25.00 per night
 "president apartment" - BGN 35.00 per night
Depending on the number of days he will stay in the hotel (example: 11 days = 10 nights) and the type of room he chooses, he can enjoy different discounts. The discounts are as follows:
room type less than 10 days between 10 and 15 days more than 15 days
room for one person does not use a reduction does not use a reduction does not use a reduction
apartment 30% of the final price 35% of the final price 50% of the final price
president apartment 10% of the final price 15% of the final price 20% of the final price
After the stay, Atanas' rating for the hotel's services can be positive or negative. If his assessment is positive, Atanas adds 25% of it to the price with the already deducted discount. If his assessment is negative, 10% is deducted from the price.
Entrance
The input is readable from the console and consists of three lines:
• First row - days of stay - integer in the interval [0 ... 365]
• Second row - room type - "room for one person", "apartment" or "president apartment"
• Third line - evaluation - "positive" or "negative"
Exit
One line must be printed on the console:
• The price for his stay at the hotel, formatted to the second decimal place.