Trade commissions.java
The company gives the following commissions to its merchants according to the city in which they work and the volume of sales s:
Grad 0 ≤ s ≤ 500 500 <s ≤ 1 000 1 000 <s ≤ 10 000 s> 10 000
Sofia 5% 7% 8% 12%
Varna 4.5% 7.5% 10% 13%
Plovdiv 5.5% 8% 12% 14.5%
Write a console program that reads the city name (text) and sales volume (real number) entered by the user, and calculates and displays the amount of the trade commission according to the table above. Display the result formatted to 2 digits after the decimal point. In case of invalid city or sales volume (negative number) print "error".