Converter for units of measure.java
Write a program that converts the distance between the following 3 units: mm, cm, m. Use the correspondences in the table below:
input unit output unit
1 meter (m) 1000 millimeters (mm)
1 meter (m) 100 centimeters (cm)
The input consists of three rows entered by the user:
• First line: conversion number - real number
• Second line: input unit of measure - text
• Third line: output unit (for the result) - text
Print the result of the conversion of the units of measure formatted to the third decimal place on the console.