Histogram.java
Histogram
N integers are given in the interval [1… 1000]. Of these, some percentage p1 are below 200, another percentage p2 are from 200 to 399, another percentage p3 are from 400 to 599, another percentage p4 are from 600 to 799 and the remaining p5 percent are from 800 up. Write a program that calculates and prints the percentages p1, p2, p3, p4 and p5.
Example: we have n = 20 numbers: 53, 7, 56, 180, 450, 920, 12, 7, 150, 250, 680, 2, 600, 200, 800, 799, 199, 46, 128, 65. We get the following distribution and visualization:
Entrance
On the first line of the input is the integer n (1 ≤ n ≤ 1000) - number of numbers. The next n lines contain an integer in the interval [1… 1000] - the numbers on which to calculate the histogram.
Exit
Print the histogram on the console - 5 lines, each of which contains a number between 0% and 100%, with two digits to the decimal point, for example 25.00%, 66.67%, 57.14%.