Left and right sum.java
Write a program that reads 2 * n numbers of integers submitted by the user and checks if the sum of the first n numbers (left sum) is equal to the sum of the second n numbers (right sum). In case of equality, print "Yes, sum =" + the amount; otherwise print "No, diff =" + the difference. The difference is calculated as a positive number (in absolute value).