Salary.java
Salary
A company boss notices that more and more employees are spending time on sites that distract them.
To prevent this, he introduces surprise checks on the open tabs of his employees' browsers. According to the site, various fines are imposed:
• "Facebook" -> BGN 150.
• "Instagram" -> BGN 100.
• "Reddit" -> BGN 50
Two lines are read from the console:
• Number of open tabs in the browser n - integer in the range [1 ... 10]
• Salary - a number in the interval [500 ... 1500]
Then n - the name of the website - text is read a number of times
If during the inspection the salary becomes less than or equal to BGN 0, it is written on the console
"You have lost your salary." and the program ends. Otherwise, after checking the console, the rest of the salary is displayed (to be written as an integer).