Sum of two numbers.java
Write a program that checks all possible combinations of a pair of numbers in the interval of two given numbers. The output is printed, which is the combination whose sum of numbers is equal to a given magic number. If there is no combination that matches the condition, a message is printed that it was not found.
Entrance
The input is readable from the console and consists of three lines:
• First line - beginning of the interval - integer in the interval [1 ... 999]
• Second line - end of the interval - integer in the interval [greater than the first number ... 1000]
• Third line - the magic number - an integer in the interval [1 ... 10000]
Exit
One line should be printed on the console, according to the result:
• If a combination is found whose sum of numbers is equal to the magic number
o "Combination N: {sequence number} ({first number} + {second number} = {magic number})"
• If no matching condition is found
o "{number of all combinations} combinations - neither equals {magic number}"