Train the Trainers.java
Train the Trainers
The "Train the trainers" course is coming to an end and the final evaluation is approaching. Your task is to help the jury that will evaluate the presentations by writing a program in which to calculate the average grade from the presentation of each presentation by a student, and finally the average success of all of them.
From the console of the first row is read the number of people in the jury n - an integer in the interval [1… 20]
Then the name of the presentation is read on a separate line - text
For each presentation on a new line read n - the number of marks - a real number in the interval [2.00… 6.00]
After calculating the average score for a specific presentation, it is printed on the console
 "{presentation name} - {average grade}."
After receiving the "Finish" command, the console prints "Student's final assessment is {average of all presentations}." and the program ends.
All grades must be formatted to the second decimal place.