Travel.java
Annie loves to travel and wants to visit several different destinations this year. When she chooses a destination, she will decide how much money she will need to get there and start saving. When she has saved enough, she will be able to travel.
From the console, each time the destination will be read first and the minimum budget that will be needed for the trip, a real number.
Then a few sums will be read, real numbers that Annie saves by working, and when she manages to raise enough for the trip, she will leave, and the console should read:
 "Going to {destination}!"
When she has visited all the destinations she wants, instead of a destination she will enter "End" and the program will end.